fixed remaining issues
